refactor blueprint boot-ups for celery inclusion
