--- purchasing/tasks.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-
